pyproject.toml
[project]
name = "bt-auto-dumper"
requires-python = ">=3.11"
keywords = []
license = {text = "MIT"}
readme = "README.md"
authors = [
    {name = "Backend Developers LTD"},
]
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dynamic = [
    "version",
]
dependencies = [
    "python-dotenv",
    "requests",
    "bittensor>=7.4.0",
]

[project.urls]
"Source" = "https://github.com/bactensor/bt-auto-dumper"
"Issue Tracker" = "https://github.com/bactensor/bt-auto-dumper/issues"

[project.scripts]
bt-auto-dumper = "bt_auto_dumper.__main__:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.version]
source = "scm"

[tool.pdm.dev-dependencies]
test = [
    "freezegun",
    "pytest-apiver",
    "pytest-asyncio",
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "pytest-xdist==3.6.1",
]
lint = [
    "codespell[toml]",
    "mypy>=1.8",
    "ruff",
    "types-freezegun",
    "types-python-dateutil",
    "types-requests",
]
release = [
    "towncrier",
]
dev = [
    "pytest>=8.3.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]  # allow for `import tests` in test files
target_package_name = "bt_auto_dumper"  # required by pytest-apiver

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "D",
    "E", "F", "I", "UP",
    "TCH005",
]
ignore = [
    "D100", "D105", "D107", "D200", "D202", "D203", "D205", "D212", "D400", "D401", "D415",
    "D101", "D102", "D103", "D104", # TODO remove once we have docstring for all public methods
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"test/**" = ["D", "F403", "F405"]

[tool.codespell]
skip = "*.min.js,pdm.lock"
ignore-words-list = "datas"

[tool.towncrier]
directory = "changelog.d"
filename = "CHANGELOG.md"
start_string = "<!-- towncrier release notes start -->\n"
underlines = ["", "", ""]
title_format = "## [{version}](https://github.com/bactensor/bt-auto-dumper/releases/tag/v{version}) - {project_date}"
issue_format = "[#{issue}](https://github.com/bactensor/bt-auto-dumper/issues/{issue})"

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[[tool.towncrier.type]]
directory = "deprecated"
name = "Deprecated"
showcontent = true

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "doc"
name = "Doc"
showcontent = true

[[tool.towncrier.type]]
directory = "infrastructure"
name = "Infrastructure"
showcontent = true

[tool.mypy]

[[tool.mypy.overrides]]
module = [
    "nox",
    "pytest",
    "tests.*",
]
ignore_missing_imports = true
